requirements.txt
mcp
openai
rich
python-dotenv